services/auth-service/requirements.txt
# AutoGraph v3 - Auth Service Dependencies

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
pyotp==2.9.0
qrcode[pil]==7.4.2
python3-saml==1.16.0

# Redis
redis==5.2.0
hiredis==3.0.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.0

# Monitoring
prometheus-client==0.21.0
sentry-sdk[fastapi]==2.18.0

# Utilities
python-dotenv==1.0.1

# Security - Secrets Management
cryptography==44.0.0

# Push Notifications
py-vapid==1.9.1
pywebpush==1.14.1
